Move service.py out of core, into a "module library" file
This is due to concerns that service.py assumes too much for core code.
